# contracts/openapi/argus-api.yaml
openapi: 3.1.0
info:
  title: ARGUS Platform API
  version: 0.1.0
  description: GENERATED by scripts/contracts.py export from backend/app. Do not hand-edit.
paths:
...
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: Me
      operationId: me_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
      security:
      - HTTPBearer: []
  /api/v1/sites/{site_id}/conditions:
    get:
      tags:
      - environment
      summary: Get Conditions
      operationId: get_conditions_api_v1_sites__site_id__conditions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: site_id
        in: path
        required: true
        schema:
          type: string
          title: Site Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteConditions'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - environment
      summary: Post Conditions
      operationId: post_conditions_api_v1_sites__site_id__conditions_post
      security:
      - HTTPBearer: []
      parameters:
      - name: site_id
        in: path
        required: true
        schema:
          type: string
          title: Site Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConditionsUpdate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwinEnvironment'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ConditionsUpdate:
      properties:
        conditions:
          $ref: '#/components/schemas/EnvironmentConditions'
        zone_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone Id
        source:
          type: string
          title: Source
          default: MANUAL
      type: object
      required:
      - conditions
      title: ConditionsUpdate
...
    EnvironmentConditions:
      properties:
        weather:
          type: string
          enum:
          - CLEAR
          - CLOUDY
          - RAIN
          - STORM
          - FOG
          title: Weather
          default: CLEAR
        temperature_c:
          type: number
          title: Temperature C
          default: 30.0
        rain_mm_h:
          type: number
          minimum: 0.0
          title: Rain Mm H
          default: 0.0
        visibility:
          type: string
          enum:
          - GOOD
          - MODERATE
          - POOR
          title: Visibility
          default: GOOD
        soil:
          type: string
          enum:
          - DRY
          - WET
          - MUD
          - ROCKY
          title: Soil
          default: DRY
      type: object
      title: EnvironmentConditions
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
      type: object
      required:
      - username
      - password
      title: LoginRequest
...
    MeResponse:
      properties:
        principal:
          $ref: '#/components/schemas/Principal'
        permissions:
          items:
            type: string
          type: array
          title: Permissions
      type: object
      required:
      - principal
      - permissions
      title: MeResponse
...
    Principal:
      properties:
        user_id:
          type: string
          title: User Id
        username:
          type: string
          title: Username
        role:
          type: string
          enum:
          - OPERATOR
          - SUPERVISOR_ADMIN
          title: Role
        operator_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Operator Id
        site_ids:
          items:
            type: string
          type: array
          title: Site Ids
          default: []
      type: object
      required:
      - user_id
      - username
      - role
      title: Principal
...
    SiteConditions:
      properties:
        site_id:
          type: string
          title: Site Id
        site:
          anyOf:
          - $ref: '#/components/schemas/TwinEnvironment'
          - type: 'null'
        zones:
          additionalProperties:
            $ref: '#/components/schemas/TwinEnvironment'
          type: object
          title: Zones
          default: {}
        environmental_difficulty:
          anyOf:
          - type: string
            enum:
            - LOW
            - MEDIUM
            - HIGH
          - type: 'null'
          title: Environmental Difficulty
      type: object
      required:
      - site_id
      title: SiteConditions
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
        principal:
          $ref: '#/components/schemas/Principal'
      type: object
      required:
      - access_token
      - expires_in
      - principal
      title: TokenResponse
    TwinEnvironment:
      properties:
        weather:
          type: string
          enum:
          - CLEAR
          - CLOUDY
          - RAIN
          - STORM
          - FOG
          title: Weather
          default: CLEAR
        temperature_c:
          type: number
          title: Temperature C
          default: 30.0
        rain_mm_h:
          type: number
          minimum: 0.0
          title: Rain Mm H
          default: 0.0
        visibility:
          type: string
          enum:
          - GOOD
          - MODERATE
          - POOR
          title: Visibility
          default: GOOD
        soil:
          type: string
          enum:
          - DRY
          - WET
          - MUD
          - ROCKY
          title: Soil
          default: DRY
        zone_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone Id
        as_of:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: As Of
      type: object
      title: TwinEnvironment
      description: Conditions as recorded for a site or zone (name kept for API compatibility).
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer